+++ modules/elementary_functions/src/include/Addition.hpp
@@ -24,10 +24,13 @@
 namespace Nelson {
 //=============================================================================
 NLSELEMENTARY_FUNCTIONS_IMPEXP ArrayOf
-double_plus_double(ArrayOf a, ArrayOf b);
+double_plus_double(const ArrayOf &a, const ArrayOf &b);
 //=============================================================================
 NLSELEMENTARY_FUNCTIONS_IMPEXP ArrayOf
-single_plus_single(ArrayOf a, ArrayOf b);
+single_plus_single(const ArrayOf &a, const ArrayOf &b);
+//=============================================================================
+NLSELEMENTARY_FUNCTIONS_IMPEXP ArrayOf
+stringArray_plus_stringArray(const ArrayOf& a, const ArrayOf &b);
 //=============================================================================
 } // namespace Nelson
 //=============================================================================